feat: `Application` dp 새로운 `Board2` 로직 으로 변경
