+++ src/main/kotlin/minesweeper/Application.kt
@@ -8,9 +8,9 @@ fun main() {
     val resultView = ResultView()
 
     val dimensions = Dimensions(inputView.readWidth(), inputView.readHeight())
-    val board = Board.initializeBoard(dimensions, DefaultCellProvider(inputView.readMineCount()))
+    val board = Board2(dimensions, MinePlacer(dimensions, inputView.readMineCount()).placeMines())
 
     resultView.startView()
-    board.detectMines()
+//    board.detectMines()
     resultView.drawBoard(board.draw())
 }
